backend/requirements.txt
django
djangorestframework
django-saml2-auth
drf-jwt
